test(tcp-server): improve testing capabilities
Add a test keyword argument to the previous serve (now serve/private)
that does different things when testing. For unit tests, there is no
logging and the function returns a procedure that shutdowns the
custodian for the server thread. This is an optional argument that
is false by default, so for production uses the server and logging
thread starts, runs and exits normally.